README.md
The objective of this work is to build a prediction model to predict stroke and evaluate the accuracy of the model. We will explore five different models to see which one produces reliable and repeatable results. The models are Decision Tree, Random Forest, SVM, Logistic Regression, KNN and XGBoost. Based on the prediction results of the models, the best-performing model will undergo cross-validation to assess its repeatability.